pyproject.toml
[project]
name = "rag-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastembed>=0.6.1",
    "pydantic-ai-slim[openai]>=1.58.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["eval"]
python_files = ["evaluate.py"]
